refactor: move session functions to ex_session.c
